Created rewrite_diagram.py
